+++ release/ui/icons/Icon/Icon.js
@@ -40,7 +40,7 @@ const iconsImports = {
 };
 const icons = Object.keys(iconsImports);
 function Icon(_a = {}) {
-    var { icon = 'default', size = 16 } = _a, props = tslib.__rest(_a, ["icon", "size"]);
+    var { icon = 'default', size = 14 } = _a, props = tslib.__rest(_a, ["icon", "size"]);
     return tslib.__asyncGenerator(this, arguments, function* Icon_1() {
         yield yield tslib.__await(({type:'span',props:{style:{
                 '--ui-icon-size': dom.inject(size, size => `${size}px`),