+++ release/ui/interaction/InputDate/index.es6.js
@@ -1,1 +1,1 @@
-export { InputDate } from './InputDate.es6.js';
+export { InputDate, inputDateApply } from './InputDate.es6.js';