+++ release/ui/popups/Modal/Modal.es6.js
@@ -57,6 +57,7 @@ function Modal(_a = {}) {
                 styles.root,
                 show.value && styles.show,
                 ((_a = hidden.value) === null || _a === void 0 ? void 0 : _a.value) && styles.hide,
+                !title && !subTitle && styles.noTitle,
             ]);
         }},children:[{type:'show',props:{state:title || subTitle || headButtonsLength},children:[{type:Flex,props:{padding:24,element:'header',vertical:true,align:'stretch',class:() => styles.header},children:[{type:'show',props:{state:title || headButtonsLength},children:[{type:Flex,props:{flex:true},children:[{type:Flex,props:{flex:true},children:[title || null]},{type:'show',props:{state:headButtonsLength},children:[{type:'div',props:{class:() => styles.headButtons},children:[headButtons.map(id => ({type:'button',props:{...buttonProps[id],'data-button':id,onclick:() => handleClose(id),class:() => styles.headButton},children:[slots[`button-${id}`] || id]}))]}]}]}]},{type:'show',props:{state:subTitle},children:[{type:'div',props:{class:() => styles.subTitle},children:[subTitle]}]}]}]},content && ({type:'div',props:{class:() => styles.content},children:[content]}),' ',children,{type:'show',props:{state:buttonsLength},children:[{type:Flex,props:{reverse:true,justify:'center',wrap:true,padding:16,gap:16},children:[buttons === null || buttons === void 0 ? void 0 : buttons.map((id, index) => ({type:Button,props:{flex:true,view:index ? 'secondary' : 'primary',...buttonProps[id],'data-button':id,onclick:() => handleClose(id)},children:[slots[`button-${id}`] || id]}))]}]}]}]});
 }